apps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
asyncpg==0.29.0
celery==5.3.4
redis==5.0.1
google-cloud-tasks==2.14.1
google-cloud-secret-manager==2.16.4
google-cloud-sql==3.8.0
